Not sure how this zverse code works, but here are a few comments: 1) I don't mind the check for valid start/end values, as long as we don't suffer a performance hit. 2) Since I don't understand the code, but it has worked fine with valid parameters for a long time,   this current commit returns the code logic back to what it was before and allows for the check. 3) If the previous logic change was correct (i.e. (start+size) <= strlen(cacheBuf)), which I can't confirm without a thoughout evaluation of the rest of the code and don't have time for this right now... then the subsequent strncpy and setSize(strlen...  checks were not needed (strncpy -> memcpy, and no setSize), since we already knew from the added strlen that we wouldn't hit a null.
This change puts the logic back to the way it has been for a long time, working with proper values,
adds a check for improper start value, (size value is insignificant with strncpy)
And still doesn't explain WHY we are getting INVALID VALUES passed when an incorrect key is used.
I suspect the invalid params means we still have a problem lingering somewhere.



git-svn-id: http://crosswire.org/svn/sword/trunk@2024 bcd7d363-81e1-0310-97ec-a550e20fc99c